+++ CHANGELOG.md
@@ -16,6 +16,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Support for go1.16. Support is now only for go1.17 and go1.18 (#2917)
 
+### Fixed
+
+-  Avoid overriding explicitly set gRPC dial options in `go.opentelemetry.io/otel/exporters/otlptrace` (#2942)
+
 ## [1.7.0/0.30.0] - 2022-04-28
 
 ### Added